Move buildForm into utils
