+++ packages/smart-forms-renderer/src/index.ts
@@ -1,6 +1,5 @@
 import { questionnaireResponseStore, questionnaireStore } from './stores';
 import type { Questionnaire, QuestionnaireResponse } from 'fhir/r4';
-import { initialiseQuestionnaireResponse } from './utils/initialise';
 import { removeEmptyAnswers } from './utils/removeEmptyAnswers';
 import type { ItemToRepopulate } from './utils/repopulateItems';
 import { getItemsToRepopulate } from './utils/repopulateItems';
@@ -12,26 +11,6 @@ export * from './hooks';
 export * from './utils';
 export * from './interfaces';
 export type { ItemToRepopulate };
-
-/**
- * Build the form with an initial Questionnaire and an optional filled QuestionnaireResponse.
- * If a QuestionnaireResponse is not provided, an empty QuestionnaireResponse is set as the initial QuestionnaireResponse.
- *
- * @author Sean Fong
- */
-export async function buildForm(
-  questionnaire: Questionnaire,
-  questionnaireResponse?: QuestionnaireResponse
-): Promise<void> {
-  await questionnaireStore.getState().buildSourceQuestionnaire(questionnaire);
-
-  const initialisedQuestionnaireResponse = initialiseQuestionnaireResponse(
-    questionnaire,
-    questionnaireResponse
-  );
-  questionnaireResponseStore.getState().buildSourceResponse(initialisedQuestionnaireResponse);
-  questionnaireStore.getState().updatePopulatedProperties(initialisedQuestionnaireResponse);
-}
 
 /**
  * Destroy the form to clean up the questionnaire and questionnaireResponse stores.
